@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace DoppioGancio\MockedClient\Route;
 
+use Closure;
 use DoppioGancio\MockedClient\Route\Exception\IncompleteRoute;
 use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
@@ -14,22 +15,15 @@ use function assert;
 use function fopen;
 use function is_resource;
 
-trait Builder
+abstract class Builder
 {
-    protected ResponseFactoryInterface $responseFactory;
-    protected StreamFactoryInterface $streamFactory;
-    protected ?string $method = null;
-    protected ?string $path   = null;
+    protected string|null $method = null;
+    protected string|null $path   = null;
 
-    public function __construct(ResponseFactoryInterface $responseFactory, StreamFactoryInterface $streamFactory)
-    {
-        $this->responseFactory = $responseFactory;
-        $this->streamFactory   = $streamFactory;
-    }
-
-    public function new(): self
-    {
-        return new self($this->responseFactory, $this->streamFactory);
+    public function __construct(
+        protected readonly ResponseFactoryInterface $responseFactory,
+        protected readonly StreamFactoryInterface $streamFactory,
+    ) {
     }
 
     public function withMethod(string $method): self
@@ -63,9 +57,7 @@ trait Builder
         return $message;
     }
 
-    /**
-     * @param array<string, string|string[]> $headers
-     */
+    /** @param array<string, string|string[]> $headers */
     protected function buildResponseFromFile(string $file, int $httpStatus, array $headers): ResponseInterface
     {
         $fp = fopen($file, 'rb');
@@ -77,9 +69,7 @@ trait Builder
         return $this->addHeaders($response, $headers);
     }
 
-    /**
-     * @param array<string, string|string[]> $headers
-     */
+    /** @param array<string, string|string[]> $headers */
     protected function buildResponseFromString(string $content, int $httpStatus, array $headers): ResponseInterface
     {
         $response = $this->responseFactory
@@ -89,8 +79,12 @@ trait Builder
         return $this->addHeaders($response, $headers);
     }
 
-    protected function buildRoute(?string $method = null, ?string $path = null, ?callable $handler = null): Route
-    {
+    /** @throws IncompleteRoute */
+    protected function buildRoute(
+        string|null $method = null,
+        string|null $path = null,
+        Closure|null $handler = null,
+    ): Route {
         if ($method === null) {
             throw new IncompleteRoute('Set parameter \"method\" before to build');
         }
