@@ -4,56 +4,46 @@ declare(strict_types=1);
 
 namespace DoppioGancio\MockedClient\Route;
 
+use Closure;
+use DoppioGancio\MockedClient\Route\Exception\IncompleteRoute;
 use GuzzleHttp\Psr7\Response;
 use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\StreamFactoryInterface;
 
 use function parse_str;
 
-class ConditionalRouteBuilder
+class ConditionalRouteBuilder extends Builder
 {
-    use Builder;
-
-    private ResponseInterface $defaulResponse;
-    protected ResponseFactoryInterface $responseFactory;
-    protected StreamFactoryInterface $streamFactory;
+    private ResponseInterface|null $defaultResponse = null;
 
     /** @var ConditionalResponse[] */
     private array $responses = [];
 
-    public function __construct(ResponseFactoryInterface $responseFactory, StreamFactoryInterface $streamFactory)
+    public function new(): ConditionalRouteBuilder
     {
-        $this->responseFactory = $responseFactory;
-        $this->streamFactory   = $streamFactory;
-        $this->defaulResponse  = new Response(404);
+        return new ConditionalRouteBuilder($this->responseFactory, $this->streamFactory);
     }
 
     public function withDefaultResponse(ResponseInterface $response): self
     {
-        $this->defaulResponse = $response;
+        $this->defaultResponse = $response;
 
         return $this;
     }
 
-    /**
-     * @param array<string, mixed> $headers
-     */
+    /** @param array<string, mixed> $headers */
     public function withDefaultStringResponse(string $queryString, int $httpStatus = 200, array $headers = []): self
     {
         return $this->withDefaultResponse(
-            $this->buildResponseFromString($queryString, $httpStatus, $headers)
+            $this->buildResponseFromString($queryString, $httpStatus, $headers),
         );
     }
 
-    /**
-     * @param array<string, mixed> $headers
-     */
+    /** @param array<string, mixed> $headers */
     public function withDefaultFileResponse(string $file, int $httpStatus = 200, array $headers = []): self
     {
         return $this->withDefaultResponse(
-            $this->buildResponseFromFile($file, $httpStatus, $headers)
+            $this->buildResponseFromFile($file, $httpStatus, $headers),
         );
     }
 
@@ -62,62 +52,59 @@ class ConditionalRouteBuilder
         parse_str($queryString, $parameters);
         $this->responses[] = new ConditionalResponse(
             $parameters,
-            $response
+            $response,
         );
 
         return $this;
     }
 
-    /**
-     * @param array<string, mixed> $headers
-     */
+    /** @param array<string, mixed> $headers */
     public function withConditionalStringResponse(
         string $queryString,
         string $content,
         int $httpStatus = 200,
-        array $headers = []
+        array $headers = [],
     ): self {
         return $this->withConditionalResponse(
             $queryString,
-            $this->buildResponseFromString($content, $httpStatus, $headers)
+            $this->buildResponseFromString($content, $httpStatus, $headers),
         );
     }
 
-    /**
-     * @param array<string, mixed> $headers
-     */
+    /** @param array<string, mixed> $headers */
     public function withConditionalFileResponse(
         string $queryString,
         string $file,
         int $httpStatus = 200,
-        array $headers = []
+        array $headers = [],
     ): self {
         return $this->withConditionalResponse(
             $queryString,
-            $this->buildResponseFromFile($file, $httpStatus, $headers)
+            $this->buildResponseFromFile($file, $httpStatus, $headers),
         );
     }
 
+    /** @throws IncompleteRoute */
     public function build(): Route
     {
         return $this->buildRoute(
             $this->method,
             $this->path,
-            $this->buildHandler()
+            $this->buildHandler(),
         );
     }
 
-    private function buildHandler(): callable
+    private function buildHandler(): Closure
     {
         return function (RequestInterface $request): ResponseInterface {
             parse_str($request->getUri()->getQuery(), $requestParameters);
             foreach ($this->responses as $conditionalResponse) {
                 if ($conditionalResponse->matchAgainst($requestParameters)) {
-                    return $conditionalResponse->getResponse();
+                    return $conditionalResponse->response;
                 }
             }
 
-            return $this->defaulResponse;
+            return $this->defaultResponse ?? new Response(404);
         };
     }
 }
